Update patch set 1
Patch Set 1:

I am ok with that. The best fix would be for the linker to notice and either fix the situation there or emit a compiler error/warning. But given that deRPC is deprecated and not widely used. Fixing so that it does not abort at runtime and logging the error/warning is better than leaving it as it is.

Patch-set: 1
Label: Verified=0