Update patch set 4
Patch Set 4:

I'm curious why this couldn't be done as JSNI? If the JSNI method is inlined, why wouldn't it stop in the right place? If the issue is source maps showing the breakpoint inside the JSNI method itself, we could post-process the sourcemap to do the right thing.

I mean I have no objection to the way this was done, but I fear we introduce too many magic AST nodes as it is.

Patch-set: 4
Reviewer: Gerrit User 5570 <5570@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0