Update patch set 4
Patch Set 4:

Yes, it was just to get the breakpoint in the right place. I hadn't thought of modifying the sourcemap. Maybe a way to generalize that would be to have a @JsInline(callersource=true) annotation on a JSNI method that causes it to be inlined and inherit the SourceInfo of of the caller? Not sure if we would use it anywhere else, though.

BTW, I have another CL to change assertions from:

x || Exceptions.throwAssertionError(message)

To:

if (!x) {
  debugger;
  throw new AssertionError(message);
}

That was what I was originally trying to do but I decided that GWT.debugger() is simpler. Any thoughts on that?

Patch-set: 4
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0