Reintroduce the performance changes to CompilationUnitInvalidator with fixes.
The prior optimization falsely assumed that the same units would not appear in
both 'units' and 'validClasses', resulting in problems where not enough units
would be invalidated.  Now, checking is done that will detect the cases where
'units' and 'validClasses' overlap, and units from the overlapping set will 
potentially be invalidated.

Change-Id: Id132950cb3ff7aa5278e12af478b2de7a7efdee1