Fix for error caused when two variables at different scopes had the same name.
The GWT compiler does not represent explicitly block scopes but allows
two different local variables at different block scopes have the same name.

It is assumed that variables with the same name can be coalesced when
lowering to the JavaScript AST representation.

The copy propagation in the DataflowOptimizer does not uphold this assumption. It
is not clear whether another optimization pass violates this assumption or not;
hence the fixup is done prior to JavaScript AST generation.

Fixes issue 7045.

Change-Id: I1d6faae19dc228a3e8ef4c59bebf4297f04b5413