Internal compiler error due to casts on lhs of assignments.
Even though casts on lhs of assignments are not valid in Java source code, there
are cases where the compiler inserts them in the AST as when dealing with generics
and erasure.

The FixAssignmentToUnbox (now FixAssignmentsToUnboxOrCast) pass was responsible for
their fixup (as per comment in GwtAstBuilder) but did not cover all possible cases.

Fixes issue 6471.

Change-Id: Iae3addd29bb075a9a96f1099b6bd02d2518294f9