Super Dev Mode: GWT.log() prints basic stack traces
For now, it's just a basic stack trace with the method names from
Throwable.getStackTrace(). Perhaps later we can do more.

In addition, use console.error to print the stack trace if available. This should
print the JavaScript stack trace leading to the GWT.log() call, which can be
convenient for jumping to source code in the debugger.

Start on issue 7460

Change-Id: I36b8bc1736c472f5545cbd33d7dbff8d293fc0c2