Primitive Java arrays (except long) backed by Typed Array in supported browsers
This patch is still in progress, and is uploaded for initial review, not to
indicate that it is complete.

The basis of this is to modify com.google.gwt.lang.Array to build typed array
instances if a) they are supported, and b) the array object is a supported
primitive (i.e. any primitive but long). This presently cheats by calling on
com.google.gwt.typedarrays.shared.TypedArrays to check for support and to build
the instances themselves - I'm unsure how to do this in a way that doesn't
require Core depending on TypedArrays, without re-declaring Array in the
TypedArrays module, or re-implementing compile- and run-time checks in Core.

It turns out that a typed array object can't have its other properties modified,
the changes need to happen to the prototype. Since we want to call instance
methods (which for normal arrays are applied via ExpandoWrapper.wrapArray) and
have a type token for instanceof (applied via Arrays.setClass and
Util.setCastableTypeMap). The current impl of this re-applies these aspects to
the prototype object each time a new array is created - this should instead just
be done once up front, but is kept simple for now to be reviewed.

Finally, several cases where it used to be assumed that the current object was
an Array, we are now referencing the methods in Array.prototype. In one case
(sorting a subset of an array), the subarray method is used. In this case, as
above, we must test if we are using typedarrays - if the TypedArrays ends up
needing a super source to clobber Arrays, we'll need one here too.

Change-Id: Ia5088ca3511a8cf9b16cc8e1d0a0fb37e06674eb