DataflowOptimizer might hang or produce incorrect results.
The DataflowOptimizer implements a lattice based composite dataflow analysis as
described in "Composing Dataflow Analysis and Transformations" by Lerner et. al
(POPL 2002).

The actual flow function is the result of applying the declared flow functions
to a possibly modified control flow graph. This resulting flow function may be
non monotonic hence compromising the termination guarantee and the fact that the
fixpoint is a solution to the dataflow equations (see Section 4.3 of the
aforementioned paper).

Fixes issue 7857.

Change-Id: I481511b549ab51ec444339361b42e53e0ddd177b