Update patch set 2
Patch Set 2: Code-Review+1

(2 comments)

Thanks Julien for doing this. A couple of things:

- missing tests for year / month movement, take a look at DatePickerTest
- Brian suggested implementing this in a different class to make sure that we do not break any old code, but I think at first trying here and see if we actually break is much better, but we should get brians opinion on this.

I really appreciate that you also covered all the formatting nits we had in the file, but if there are so many its much easier to send a reformatting change upfront, so the actual review is easier to do.

Patch-set: 2
Label: Code-Review=+1