Update patch set 2
Patch Set 2:

Hi Daniel,

I missed the DatePickerTest. I will update it.

Concerning the Brian comment, it was related to the old patch that was never submitted. This patch is a new one that shared nothing with the old one. I especially pay attention to keep backward compatibility. So If you don't use any new options, the date picker is rendered as it was before (except that now it uses single quotation mark instead of double quotation mark for month navigation. But I can change that if you want)

The formatting nits were automatically fixed by intellij, I will check my config to avoid that in the future

Patch-set: 2
Label: Verified=0