Retain empty stubs for classes that are @GwtIncompatible.
A @GwtIncompatible class is replaced by the compiler by an empty stub
class. The stub class is a final class with a private default constructor
to ensure it is not extended nor instantiated.

Previous to this fix, a @GwtIncompatible top level class would cause a
compiler error (due to attempting to compile an empty unit).

Change-Id: Id2fec62e005a3f90df6cb5de9d56203018f991f3