Updated ASM 3.1 to 4.1 to support Java7 class file format.
ASM is currently rebased into com.google.gwt.dev.asm and there is
only one small modification to it in (MethodWriter.java:1201).

The reason for the modification seems to be that emma (the code
coverage tool used with GWT) sometimes produces classfiles that
violate the class file spec, yet they are successfully loaded by
the JVM. This "incorrect" class files when processed by ASM result
in class files that can NOT be loaded by the JVM.

Due to the stucture of ASM it is difficult to externalize by
subclassing and intercepting. Here it would be best to either (1) move
to code coverage in javascript, or (2) patch emma instead.

Change-Id: I5821dab924587a186f2d5e82408de5cd0af66ee6