Update patch set 3
Patch Set 3:

@John, 

I am not too familiar with ASM either. There is probably a way to fix by writing a filter visitor. The current fix relies on access to protected fields that don't seem to be have a public interface. 

I checked at the EMMA site, and there is a patch there that should solve the issue. 

Are there any other reasons to have ASM rebased?

Patch-set: 3
Label: Verified=0