Update patch set 3
Patch Set 3:

We used to have a lot more local changes, but they have gradually gone away with newer versions, so if this is the last one and a patch to emma fixes it, then I think we can do away with having a locally modified copy.  There are still things that get jarjar'd, and I don't know if we care about not stepping on other versions of ASM that the app wants to use.  My guess is we could do away with the rebased version, but it would take some testing.

Patch-set: 3
Reviewer: Gerrit User 5100 <5100@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0