Workaround JDT bug 397462 which causes a compile error when there is none.
This is a temporary patch to turn some JDT errors into warnings in
somewhat narrow cases.

The following example code fails to compile without this fix:
  public class Base {
    static Base factoryMethod() {
      return null;
    }
   }
  public class Child<S> extends Base {
    static <S> Child<S>  factoryMethod() {
      return null;
    }
   }

JLS 8.4.2 is explicit that this case is not a clash even though Base.factoryMethod() and
Child.factoryMethod() have diferent number of type paramters.

A more interesting case is when both have type parameters but a different number of them
as in (which fails to compile without this fix):

  public class Base<T> {
    static <T> Base<T> factoryMethod() {
      return null;
    }
   }
  public class Child<T,S> extends Base<T> {
    static <T,S> Child<T,S>  factoryMethod() {
      return null;
    }
   }

As per JLS 8.4.2 rules these two constitute a name clash as the two signatures have a
different number of type parameters but none is the erasure of the other.
Interestingly javac does not mark this case as an error and there are valid use cases
as the one in the example.

This patch relaxes the rules in the case of static functions and emits warnings
instead of errors.

Change-Id: I3767690ce82e0763224c27c9d0e169034d857a03