Update patch set 7
Patch Set 7:

Ah, my bad, I ran it from Eclipse and forgot to configure Ant (re. the ValidationTool error).

As for the error we see in Jenkins, I think we have to update the requestfactory-apt.jar from GWT_TOOLS to a newer version (it's 2 years old!).

Either that or change the build.xml to use the requestfactory-apt.jar we're just generating (not quite possible currently, given how our build scripts are made), or disable annotation processing in user/ and don't run RF tests there but only from requestfactory/

I think it's easier to update the requestfactory-apt.jar in GWT_TOOLS ;-)

Patch-set: 7