Update patch set 7
Patch Set 7:

Looking more closing, I'm not sure it'll be enough: JsonRpcService don't have a mapping to a "domain class", so the ValidationTool must be used (rather than the annotation processor at compile-time) with the "-client" flag.

Ideally, we'd fix the annotation processor so it doesn't require domain mapping for JsonRpcService.

Let's revert that change for now.

Patch-set: 7