Update patch set 7
Patch Set 7:

An easy *workaround* is to annotate the interface with @SkipInterfaceValidation (in addition to passing it to the ValidationTool in requestfactory/build.xml).

See https://gwt-review.googlesource.com/4280

Tried it with "ant testrf". We'd still want to fix the annotation processor, but I think it's worth rewrite so maybe it can wait…

Patch-set: 7