Update patch set 1
Patch Set 1:

Noticed this looking for deprecated code.  It seems that if you have JSNI code that tries to reference (e.g.) "@int::class", you actually end up with boolean.class.

Not sure why anyone would want to access class literals in JSNI code like this, and there aren't any uses of it within Google, but seems worth fixing.

I can demonstrate the bug in a GWTTestCase with:

  public void testClassJSNI() {
    Assert.assertEquals("boolean", getBooleanClass().getName());
    Assert.assertEquals("int", getIntClass().getName());
  }

  private static native Class<?> getBooleanClass() /*-{
    return @boolean::class;
  }-*/;

  private static native Class<?> getIntClass() /*-{
    return @int::class;
  }-*/;

but not sure where to add it.

Patch-set: 1
Label: Verified=0