Update patch set 1
Patch Set 1:

Is this going away *entirely* in 2.6, or being disabled by default? Dropping it in 3.0 still lets the 2.x line be maintained as needed (by those who are doing the needing), but cutting it out in 2.x in this way makes it very hard to maintain a tool/app that can take advantage of the rest of the 2.6 release while still working in the default browsers of both Windows XP and Vista.

Patch-set: 1
Label: Verified=0