Update patch set 6
Patch Set 6:

I think there are advantages to have multiple patches instead of one giant patch and reviewed by individuals who are familiar with that specific part and can be reverted if anything is broken unintentionally. Also perhaps some specific parts better to be simplified instead of just deleting some code when IE6/7 is out of the picture.

Also dropping the support by first dropping the permutation in 2.6 (so that people can easily revive if they need it) and then start to delete the code after 2.6 might be a better approach. I thought that was what we discussed earlier?

Patch-set: 6