Fixed undefined class literal field JavaScript exception in CodeSplitter.
CodeSplitter uses fixup passes to ensure that load dependencies are satisfied.

Load dependencies are dependencies that are not taken into account by the control flow
analysis. Among those are the requirement that the "defineSeed" for a superclass must
always be declared before that of a subclass. These dependencies occur due to the
fact that the code that sets up class hierarchy (and more) runs when the script is
loaded.

The fixups where incorrectly handling several cases:

(1) Class literal field declaration refers to the superclass class literal field.
This case was only partially handled, only class literal fields that had an
explicit class literal reference where fixed (that is enough for fully
optimized code).

(2) Fields whose initialization include a string (e.g. by a method parameter).
Only fields that where assigned a fragment correctly fix up the string if it was in
a different exclusive fragment.

Change-Id: Ia2c4367158878383725470588e96d0decc83e6ce