Update patch set 2
Patch Set 2:

I discovered that UserAgentProperty will not be even generated if the runtime user agent assertion is disabled - this is uncommon for unit tests but we should not rely on the flag. I'll first push a patch for making UserAgent public and flag independent then return back to this one.

Patch-set: 2