Update patch set 1
Patch Set 1:

Added also Brian.

I don't see anything wrong with a generated returning the class itself. It is very likely that gwt-exporter already uses that (if not anything already uses it in SDK).
On the other side, we don't have much options as config properties are not usable for deferred-binding. The reasons for that are not obvious to me but it doesn't look like trivial to change at the moment.

Patch-set: 1