Update patch set 1
Patch Set 1:

I think it is already verbose because we needed to introduce a generator that could have been a simple xml config if it wasn't a configuration property; adding also another additional class for the same reasons is kind of unfortunate.

I'm not much concerned about the surprise factor as generator is quite trivial and what is going is documented in the class javadoc but if you think it would be better, I can introduce the extra class.

Patch-set: 1