Update patch set 1
Patch Set 1:

Yes, that makes sense. And that is something we could easily introduce using <evaluate-with> proposal that we discussed during GWT.create.

On the other hand, there is pretty good alternative out in the wild. Not as convenient as the magic GWT method but I like the idea:
  https://github.com/Legioth/PropertySource

Patch-set: 1