Clean up and refactor CodeSplitter & CodeSplitter2
Fixes random issues in CodeSplitter2 due to the fact that the fixup pass
for load order dependencies was done incorrectly; it was done when deciding
how to merge but was not taken into account when computing the atoms the
resulting fragments.

This patch unifies the logic with the original CodeSplitter, making splitting
a two stage process: (1) decide how to group splitpoints into fragments, and
(2) compute how the code is split into fragments taking into account load
order dependencies.

Change-Id: I45f27fd83263fabfa2e05c45e9e01eb79e8071f9