Update patch set 2
Patch Set 2:

I am not sure of all the implications. Before the patch Long values were Java Objects, after the patch they are JavaScript objects.

The seems to be an assumption that the obfuscator will not rename the (h,l.m) fields. But I am completely unfamiliar with the code.

I'll forward the review to cromwellian.

Patch-set: 2