Update patch set 1
Patch Set 1:

(2 comments)

This solution is much better than the old way of writing out the user agent, but I feel that adding code inside of the gwt.xml is not that great. It works okay for small selection scrips and looks okay so far.
Absolutely not sure about this:
What do you think instead of adding the code directly in the gwt.xml we add an entry to a file (like we do in linkers) to look for the javascript. This way you can write much bigger scripts nicer. We could even have a config property for a common file that contains utility functions.
Might be overkill though.

Patch-set: 1
Reviewer: Gerrit User 5385 <5385@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0