Update patch set 1
Patch Set 1:

Well, yes and no. We're still compatible with Java 6 in the sense that everything that parsed OK in Java 6 still parse OK in Java 7; we're incompatible in that "+1" failed in Java 6 and now pass in Java 7.

What it means is that prod mode would behave as Java 7 even when you pass "-sourceLevel 6" to the GWT compiler. But that's not really different from JVMs actually: you can compile classes with "-source 6 -target 6" but the behavior of Integer.parseInt() will depend on the JVM you *run* that code in. With GWT, the "runtime" would be "Java 7" after that change.

I think we should put it in the GWT 2.6 release notes (not necessarily the announcement), something like:

 GWT now supports Java 7 sources, and the compiled JS will behave as if it were a Java 7 runtime (i.e. Integer.parseInt("+1") no longer throws)
 You can pass -sourceLevel 6 to the compiler to treat sources as Java 6, but the behavior of compiled code will stay the same (re. Integer.parseInt notably).

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0