CLEANUP: Replace c.g.g.dev.util.colllect by Guava collections.
The rationale of this change is to cleanup the codebase and use off
the shelf, well maintained implementations where possible.

The collections implementations, while saving some memory, introduced
unwanted complexity and had the potential for introducing exceptions
in some cases. Operations that modify the contents these collections
need to be called through static methods (to, for example avoid an
exception when adding to the empty list) a practice that is not uniformily
enforced in the compiler.

The big GWT applications that hit the memory barrier can switch to use
a 64-bit JVM.

The real solution to the memory problem, IMHO, is to redesign the
global optimization pathway so that it operates on a smaller abstraction
without the need to traverse the whole AST.

Change-Id: Ifa1a760086e8a73ee78879d6303e1f25edd67fcd