Update patch set 1
Patch Set 1:

Be very careful about doing this -- the reason they were done this way is so AdWords could compile.  Using these custom implementations saved a pretty substantial amount of memory in the AST.

You should look at the size of the heap with and without this change (you will need at least 16G to run jhat on it when compiling AdWords).

Patch-set: 1