Update patch set 2
Patch Set 2: Code-Review+1

Do we really need to use a rebased version of Guava? Couldn't you depend on the Guava "trunk" inside Google? I don't think we use any @Beta API, and if anything were @Deprecated at some point in time, we'd have a 18 months timeframe to update; and Guava is never surfaced in our API so we shouldn't ever run the risk of a breaking change due to Guava removing an API.

The same could be said for our other dependencies (protobuf and closure compiler). Or is this common practice within Google to get "stability" that way? (snapshotting an API and rebasing it; i.e. fork it)

As far as this change is concerned though, +1, let's update.

Patch-set: 2
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+1