Update patch set 1
Patch Set 1: Code-Review-1

You should probably rebase that change on top of the "move to Guava" one. And until you've proven that memory is not an issue, you should probably put a "-1" or "-2" on that other change to make sure nobody will merge it too early (e.g. before you test it with AdWords).

As it is, -1 (for the reasons Brian gave). I'll remove it when it's rebased or when the "move to Guava" has been merged.

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=-1