Update patch set 1
Patch Set 1: Code-Review+1

I'm a bit concerned that some people might be relying on the current behavior (e.g. you edit a cell with an "invalid" value, if you hit the "escape" key you clearly intend to cancel the edit, but if you hit the "enter" key (without changing the value) you're implying "that's the value I want here", and the app can bug you with an alert saying "this value is invalid, it'll break this and that".

Clearly not the best UX, but I believe there must be people doing it that way.

With that change, when you edit the invalid value, then whatever you're doing to exit the field (without actually changing the value), the popup won't show.

Matthew, what do you think? I'm about to give a +2 here but I'd like a second opinion on that change in behavior.

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+1