Update patch set 1
Patch Set 1:

I'm not familiar with the code but from Thomas' explanation, it doesn't look like the behavior will be correct for invalid values and the user will never see a validation error unless he/she changes the value. Doesn't sound good.

On the other hand, if the field started with an invalid value and the user never edited the field, then the invalid value will potentially end up in the server anyway if there are no secondary checks. So it was already a problem or they already solved it.

So I'm fine with the change if it looks like the only potential negative impact.

Patch-set: 1
Label: Code-Review=0