Update patch set 2
Patch Set 2:

(2 comments)

I agree that a fully functional fake would be useful, but that's a different debate and the fake would be much more complicated. In practice, Java tests can use mocks to expect and capture the scheduled commands, and GWT tests can call SchedulerImpl.flushEntry/FinallyCommands() to execute the commands synchronously.

Here I'm only trying to make the current stub easier to use: the only sensible things to with the stub is checking that getRepeating/ScheduledCommands() return empty lists, or expecting them not to be empty and execute the scheduled commands. executeCommands() simplifies that. Does it make sense?

Patch-set: 2