Update patch set 2
Patch Set 2: Code-Review+2

(38 comments)

Seems fine assuming that performance tests show it's not a regression.

Thanks for doing this, but wow that's a lot to review. I made a bunch of comments on further work, but I think it should all be in a separate CL since I don't want to go through this again.

Patch-set: 2