Update patch set 2
Patch Set 2:

Update:

The bug in the HtmlUnit that was causing TypedArray tests in web mode is fixed yesterday and they are planning to do a new release (2.13) soon. I am planning to wait for the new release before submission.

I also created another patch (4530) that is fixing TypedArray tests in dev mode but it is still puzzling to me why HtmlUnit dev-mode utilizes a different code path then the firefox dev-mode. I'm investigating that before submitting the fix.

Patch-set: 2
Label: Verified=0