Update patch set 2
Patch Set 2:

So the root cause of the problem is htmlunit's JavaScriptEngine.callFunction returns Long which was never happened before and Long is not really a valid js type. So I filed a bug to HtmlUnit.

In the meantime, I'll try remove the buggy long and float support from the dev mode protocol so that we will fail-fast if something like this ever happens again.

If htmlunit guys doesn't recognize this as a bug, we can discuss the alternatives.

[1] https://sourceforge.net/p/htmlunit/bugs/1543/

Patch-set: 2