Update patch set 6
Patch Set 6:

You're right, tests aren't run by Ant. Someone will have to fix the build.xml (though we'll be migrating away from Ant in the near future, so it might –or might not– be worth the effort).

Reason might be that a) tests failed in DevMode, and/or b) the code expects a standards-compliant modern browser, and our current version of HtmlUnit wouldn't support many/some of those tests.

Patch-set: 6