Adds conversion logic for non-js numeric types to dev-mode compatible types.
With the HtmlUnit upgrade I discovered that htmlunit js engine can return
types that are not available in javascript. This caused htmlunit devmode
to utilize the buggy codepath for transerring longs in BrowserChannel.

This change ensures that we are sending numeric types as either Integer or
Double to better mimic other devmodes.

Change-Id: Ic58b6c08a6c56aba455246cb8615e17c162d61d7
Review-Link: https://gwt-review.googlesource.com/#/c/4592/