Removes the code from BrowserChannel for unused numeric types.
Devmode doesn't use float and long on the wire. The code path for both
types are untested and actually contains bugs in both C++ and Java sides.
This patch marks these types as unused and removes related code from
BrowserChannel to prevent accidental use - which has happened after an
upgrade of htmlunit.

Change-Id: I27131f2f0a1f772f6f17c4296e486d6e7208a1f2