Change the generated code in assertions for easier debugging  - add a "debugger" statement before throwing  - do the throw from the caller for one less frame in the stack trace
Unfortunately the stack trace doesn't get properly logged to the JavaScript
console due to a catch() clause in Impl.entry(). However, changing that seems
risky enough for it to be a separate CL.

(The magic methods in "Exceptions" seem kind of silly now. I assume
they're needed because constructors are somehow harder to look up.)

Change-Id: I655d2244a01d97ea6a538f695a85cc3eb3a1cb5b