Removes the extra precision from date when entered from picker or textbox
When thedate  entered from text box or date picker, the date might have higher
precision then the datebox can display. This causes an extra value change event
when the textbox loses focus or date picker closes - which happens any time a date
is selected from the date picker. So it ends up in two value change event fires.

The change makes a round-trip on format and parse while accepting the date from
user to ensure that precison is no higher than the precision that can be
represented in the datebox.

fixes issue 4785

Change-Id: I35f50d0d5239560c1466697678040d8eefedc2a4