Update patch set 1
Patch Set 1:

(1 comment)

Thanks for the review.

Testing it for all timezones and transitions doesn't look practically possible to me. For that I need to manually switch to a corresponding country in OS, restart chrome, find DST times for that country and test it.

However, the bug only occurs when creating the date with new Date(YYYY, mm, dd) and transition happens at midnight. So I was able to reproduce it in BST to see that if it is really fixed with the patch for java.util.Date. Also inspecting the code in java.util.Date#fixDaylightSavings make me fairly confident that this will not break unless we have some bugs in java.util.Date.

Patch-set: 1