Avoid naming clashes between static members and top level names.
Defining static functions and methods in the global scope has the following
implications: (1) global variables might be overwritten when assigning a static
variable or defining a static function and (2) because static variables are
assumed to be initially undefined and are not explicitly initialized, if there
is a name clash the variable might not initially undefined when it should be.

We avoid clashes by blacklisting names that are normally at the top scope.

A better solution is to ensure GWT code is always enclosed by its own scope.

Change-Id: Ifd4b3a111764c239863062b372c5a06e7f7d6ef4