Update patch set 5
Patch Set 5:

AIUI, the issue is with modules with a rename-to="ui" (where 'ui' can be anything that's likely to clash with an obfuscated variable name). Given how GWT assigns/allocates variable names (IIRC), it's likely to use any combination of lowercase and uppercase US-ASCII alphanum of length 1, 2 or 3 chars (e.g. "ui", or "AN" [1]), which can easily conflict with those rename-to="" with short names (the way the built-in linkers work at least).

This is very much outside my area of expertise, and I really don't understand the code from this patch, but thought that it might be related to issue 8229 (talking about variable name clashes), either fixing it or providing a "framework" to make it easier to fix (as I said, I don't understand that code :) )

[1] https://groups.google.com/d/msg/google-web-toolkit/aQYWWAGqfJI/LUF98UkRRuQJ

Patch-set: 5