Change ArrayList.removeRange to not throw IOBE when from == to == 0
This is confusing behavior, but is an attempt to match Java's ArrayList.

The Javadoc says "(If toIndex==fromIndex, this operation has no effect.)"

Java in the wild seems to fail if toIndex == fromIndex == this.size() != 0,
so we match this confusing behavior.

To test the behavior on your java, compile and run the following:

import java.util.ArrayList;
public class a extends ArrayList{
  private static class Stack extends ArrayList {
    public Stack(int i) { super(i); }
    public void r(int from, int to) { removeRange(from, to);  }
  }
  public static void main(String[] args) {
    Stack s = new Stack(1);
    s.r(0,0); System.out.println("0,0 success");
    s.r(1,1); System.out.println("1,1 success");
  }
}


Change-Id: Ifb5d5998d8ddcab1a126107672e9a71f8b760fb9
Review-Link: https://gwt-review.googlesource.com/#/c/4731/