Update patch set 6
Patch Set 6:

I guess the cover message is a bit confusing. It describes how it was implemented before the fix.

With this patch user variables that start with $t are no longer a problem.

FWIW there is an upcoming patch that does away with this way of naming temporary.

Patch-set: 6
Reviewer: Gerrit User 5105 <5105@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0