currently CompilePerms writes a PermutationResult instance to disk by sticking it into a FileBackedObject instance. later during Link the PermutationResult is retrieved from disk using that same FileBackedObject instance. and this is the path way by which compiled javascript source and its description (statement ranges) reach the linker.
for separate compilation something similar needs to occur but the
reading and writing needs to be into and out of a library zip file.

this patch adds a ZipEntryBackedObject class and pulls out and uses
a common interface it shares with FileBackedObject. this patch does
not yet make use of the new class, that will come later in library
support and Compiler entry point change patches.

Change-Id: I26c674af85c184a9b467659997550fc0f51e8977