Ensure int values for pixels.
Browsers started to send sub pixel values for things like
element.getBoundingBox().top.

Added methods that return double to DOMImpl and Element
and modified existing ones to use it with a fast round (e.g. width | 0)

This patch does not touch IE6 and IE8 DOM impl since this is not an issue
with those two.

Change-Id: I32f1eb1d9b8591d71d3662455c57bfd4ad696661