Update patch set 2
Patch Set 2:

I could not make eclipse build the annotation on core/test/com/google/gwt/dev/javac/typemodel/test/package-info.java.

It was always displaying an error about the value not being present on the annotation (which is not true). As soon as you changed the file and triggered an incremental compile, the error was gone.

I simply removed that file from the eclipse build as well as the Public.java file without a package.

With those changes gwt-dev and gwt-user are building fine for me in eclipse.

Patch-set: 2
Reviewer: Gerrit User 5385 <5385@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0