Update patch set 2
Patch Set 2:

Yes, that's true. Nevertheless, I've made this change locally on multiple occasions, and I just heard from another user who spent hours tracking down a serialization problem when this change would have pointed him directly to the class responsible. (He had had to write a test, which was executed unobfuscated, but that was a small part of the investigation.) I grant that my situation is probably unusual: Guava has a lot of internal implementation types, so if you've got an ImmutableList, you might have a SingletonImmutableList, a RegularImmutableList, some kind of ImmutableAsList, or a slightly different set of types if you're on the client. Figuring out exactly which one you're dealing with is half the battle, especially if it's a remote part of some object graph, possibly with the problem occurring sporadically depending on exactly which code path creates the list.

I see that there's a check mark in the Code-Review column in Gerrit, so I'm not actually sure whether I need to be making this pitch at this point or not :) Let me know if the pitch was unconvincing or if there's anything else I need to do.

Patch-set: 2
Reviewer: Gerrit User 6141 <6141@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0