Fix SplitLayoutPanel's resizing in a scrolled browser window.
Without this change, resizing the north panel works correctly
regardless of the window scroll, but the south panel will "jump" up
when you start to drag it. Adding instead of subtracting the offset
(mouse position minus splitter position) seems to fix it.

This only affects "reversed" splitters when the main browser window
has been scrolled -- so, south panels if the window is scrolled
vertically, or east if the window's scrolled horizontally.

It looks like this was reported before, although it was a while ago:
https://code.google.com/p/google-web-toolkit/issues/detail?id=4755

Change-Id: I055ffff82e73823059d5c19ab1e7aeb3de807c57
Review-Link: https://gwt-review.googlesource.com/#/c/4982/