Adds a unit cache that finds compilation unit instances in a library group.
ContentId based lookups are not supported since having multiple versions
of the same compilation unit in a set of libraries would be an error.

Removals are not supported since libraries are immutable and the
removals would only be needed if library contents were invalid.

This class is an integral part of reusing previous work (previously
compiled compilation units) from library files created in previous
compiles of dependency library modules.

Change-Id: I9b67f8fecbcad56a0b1197532fccc811ff3030c1