Add an option to set the port used in GWTTestCases
In manual mode the port is set randomly thus invalidating all breakpoints
that might have been set to debug an issue in some browser.
This change lets the developer specify a port on the command line, so that
test can easily be executed again without loosing all context in a browsers
debugger.

Change-Id: Ia6e28148e2e397780453f2d448fd58880968e7d2