Update patch set 10
Patch Set 10:

I was able to repro the RequestFactorySuite failures and have a very straightforward fix. Should I submit it as a separate  CR since this one has been merged?

With the fix there's only one test in RequestFactoryJreSuite that fails, RequestFactoryUnicodeEscapingJreTest.testServerToClientBMP, but the same test passes in RequestFactorySuite. I'm going to ignore this failure and treat it as an issue with my setup as it fails for me locally when I run against the master branch commit from before my change was merged.

Patch-set: 10