Update patch set 4
Patch Set 4:

FYI, I'm not sure this is a good idea in general.

Without adding any delay to auto-close the submenu, this usually ends up in very poor user experience and makes it harder to navigate the menu. (I actually hate when the websites do that.)

I would much have preferred if there would exist a delay before switching the menu items. I'm not going to "-1" because this is optional and can be fixed without a breaking change. I'm leaving it to other reviewers to decide.

Patch-set: 4
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0