Update patch set 1
Patch Set 1:

Sorry, not trying to pick a fight, but I am trying to watch changes more closely after master didn't build for a week or so, and especially as we get close to the 'feature complete' date.

The change set was only up for 29 minutes before it was merged, and late on a friday evening (at least for the eastern US) - while I certainly don't doubt your or Roberto's talents, I did think that seemed to be a little shorter than usual (except perhaps for revert commits).

Even as a steering committee member who reads nearly any commit I have the chance to glance at (just under half of the last 50, according to gmail...), I only get to see the code and the commits, but not always the thinking going into the direction we're heading. As I understand it, we use Gerrit with its one-commit-per-changeset approach to keep new features/functionality atomic when they are created and merged, but as this one is incomplete, it is hard to given any insight - I necessarily missed out on any design doc or meat-space part of this conversation.

I'm not asking for or expecting oversight of every change that goes in, but trying to keep up, and see what this is - at this point, it is still dead code (you don't seem to have any other open changesets), and while I can read what it *does*, the info I have so far gives no real idea as to *why*. That's all I'm hoping to understand better as a non-Googler in future changesets.

Patch-set: 1