Refactors TypeOracleBuilder/TypeOracleMediator/subclasses
Combines TypeOracleMediator and TypeOracleMediatorFromSource since they
both were claiming to operate on CompilationUnits and no non-test code
existed that depended on the difference between them.

Replaced "Builder" and "Mediator" verbiage with the more accurate term
"Updater".

This patch is mostly just a cleanup and clarification. The only
behavior change is that where previously TypeOracleBuilder instantiated
a TypeOracle instance during its initialization, now an instance is
passed in via constructor.

Change-Id: I8c9078e651a3eab03730ad36c5db0c9dd6499216