Update patch set 1
Patch Set 1:

Hey Danilo, sorry for the delay.  Looks good to me.

Sorry to ask this so late, but is it possible for you to write a unit test for this?  E.g., change the value from "A"->"B"->"B">-"A" or whatever and make sure you only get change events for A->B and B->?

Patch-set: 1
Reviewer: Gerrit User 5000 <5000@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0