Update patch set 10
Patch Set 10:

FYI:

This was only reverted to solve a conflict for @stalcup to be able to get incremental compile into trunk.
We have two different changes about Array: One being (possibly) faster but less secure in terms of JS code breaking (this one) and we got another proposal which might be slightly slower but does not have the potential to break any thirdparty code. We will open two new reviews to see which change is better.

Patch-set: 10