Supplies CompilerContext all over the call graph.
Separate compilation creates the need to have access to loaded
precompiled libraries, the current library builder, options
and the global unitCache all over the call graph. Rather than
add four new parameters to most of the call graph it was
decided to group them together and add just one new parameter
to most of the call graph (and in some places remove some now
redundant parameters).

Going further, existing compilation assumes that build,
source, and public resources should always be accessed out of
resource oracles attached to the module object. But with
separate compilation resource loading needs to be serviced
from a combination of the current module and precompiled
libraries. So the CompilerContext also serves as a consistent
place for compiler code to access resource oracles without
the accessing code having to know from whence those resources
originated.

The rest of the separate compilation work makes heavy use of
the CompilerContext and its presence throughout the call graph,
so while it's weird to be submitting a CL that introduces so
many temporarily unused function parameters its very useful in
that it opens up a lot of the remaining separate compilation
changes to be reviewed in small pieces.

To keep things sane the CompilerContext contents are treated as
read only except at the very top of the function call tree
(like the main function and tests).

Some alternative designs were considered but rejected for
various reasons:
 * dependency injection, would require a massive general refactoring
 * singletons, not thread safe
 * thread local variables, too magical and would lead to odd
   instance hand-offs between threads

That being said design feedback and counter suggestions are welcome.

Change-Id: I1c56dd7cd5d8b333f8ffac455be280a8718d8cd1